split encoder_decoder.py
